% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_lights.R
\name{alert_light}
\alias{alert_light}
\title{Blink a light for a few seconds}
\usage{
alert_light(bridge_url, user, light_id)
}
\arguments{
\item{bridge_url}{Hue bridge URL E.g. "http://192.168.1.100"}

\item{user}{user id, to be obtained by \code{create_user}}

\item{light_id}{used to address a specific light. Can be looked up by \code{dump_lights_data}}
}
\description{
Blink a light for a few seconds
}
