% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_colors.R
\name{convert_xy}
\alias{convert_xy}
\title{Convert a CSS color to Hue XY colorspace}
\usage{
convert_xy(color)
}
\arguments{
\item{color}{A CSS hex color definition, e.g. \code{"#FF0000"}}
}
\value{
A numeric vector of length 2, defining a color in XY colorspace
}
\description{
Convert a CSS color to Hue XY colorspace
}
\examples{
convert_xy("#FF0000")
}
