% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{create_user}
\alias{create_user}
\title{Register a new user with a hue bridge.}
\usage{
create_user(bridge_url, app_name, device_name)
}
\arguments{
\item{bridge_url}{Hue bridge URL E.g. "http://192.168.1.100"}

\item{app_name}{Arbitrary name (See Hue API Documentation)}

\item{device_name}{Arbitrary name (See Hue API Documentation)}
}
\value{
A new username (character)
}
\description{
\strong{Make sure the bridge's button was pressed no longer than 30s prior to calling this function!}
}
\examples{
\dontrun{
create_user("http://192.168.1.100", "rhue", "test")
}
}
