% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{dump_lights_data}
\alias{dump_lights_data}
\title{Dump available data about all lights paired with the bridge.}
\usage{
dump_lights_data(bridge_url, user)
}
\arguments{
\item{bridge_url}{Hue bridge URL E.g. "http://192.168.1.100"}

\item{user}{user id, to be obtained by \code{create_user}}
}
\value{
A list of data.frames containing metadata. One for each type of light in use.
}
\description{
Dump available data about all lights paired with the bridge.
}
