% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_light_data}
\alias{get_light_data}
\title{Get state and metadata of any light paired with the bridge.}
\usage{
get_light_data(bridge_url, user, light_id)
}
\arguments{
\item{bridge_url}{Hue bridge URL E.g. "http://192.168.1.100"}

\item{user}{user id, to be obtained by \code{create_user}}

\item{light_id}{used to address a specific light. Can be looked up by \code{dump_lights_data},}
}
\value{
A list containing state and metadata of a hue light.
}
\description{
Get state and metadata of any light paired with the bridge.
}
