% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_lights.R
\name{set_light_color}
\alias{set_light_color}
\title{Change light color}
\usage{
set_light_color(bridge_url, user, light_id, color)
}
\arguments{
\item{bridge_url}{Hue bridge URL E.g. "http://192.168.1.100"}

\item{user}{user id, to be obtained by \code{create_user}}

\item{light_id}{used to address a specific light. Can be looked up by \code{dump_lights_data}}

\item{color}{hex-encoded color, e.g. "#FF0000"}
}
\description{
This function does not check if the light actually has the capability to change color,
Lamp metadata, obtainable by \code{get_light_data} should contain that information.
}
